Updating People and Modules
The list was out of date, here are some proposed changes. Hopefully they are correct.

I'm also proposing to remove the list of people that's automatically pulled from GitHub data.
It's clearly not representative of who is actually working on the project.

I've also done a couple minor updates to links to the doc.